requirements.txt
google-generativeai
python-dotenv==1.0.1
tkinterdnd2==0.4.3 
streamlit